备份 202003012更新说明（大更新） 一、服务端和后台前端 1.增加商品推荐字段，为商品订单增加独立物流字段（用在新增的订单取消和未来的退货功能） 2.在后台首页增加两个图表：销售额最高的商品排名，各产品栏目的销售量饼状图 3.订单图表和产品销量的统计条件变更 说明：原来的计算方式是所有订单都计算进去，现在是有效订单才记录（作废、退货、取消的订单不列入计算） 4.增加功能：增加订单作废功能（管理员操作） （1）订单作废后，将把订单中购买的商品数量返回给相应的商品 （2）只有满足系统规定的订单才能作废 5.增加功能：增加订单取消功能 （1）首先由客户申请取消订单，只有满足系统规定的订单才能申请取消。 备注：如果是网上支付且未付款，则客户可以直接取消，无需管理员处理 （2）管理员对客户申请的订单进行处理，如果允许取消，则直接取消订单；如果拒绝取消，则订单恢复到申请取消前的状态。 说明：订单取消后，将把订单中购买的商品数量返回给相应的商品 6.改进功能：自动购买程序将只会选择已上架的商品，原来是选择所有商品，可能造成不必要的购买失败（失败了也不会影响购买程序，但不合理） 7.修改BUG：后台的修改订单界面，修改订单的页面将符合系统要求。 备注：原界面只允许未付款订单修改
未来发展：
1.对异常类进行划分，除了web异常，还应该增加权限异常等（后续根据实际要求确认）
2.redis跨域登陆，增加登陆的性能
3.订单模块
4.引入node.js解决跨域访问问题
5.vue的全新前端架构：目标动态菜单

+++ src/main/java/com/lcvc/ebuy_springboot/service/impl/AdminRoleServiceImpl.java
@@ -118,6 +118,10 @@ public class AdminRoleServiceImpl implements AdminRoleService {
         if(roleIds.length==0){
             throw new MyWebException("操作失败：请选择要移除的角色");
         }
+        List<AdminRole> adminRoles=adminRoleDao.getAdminRoleByAdminId(adminId);//获取该角色的所有集合
+        if(roleIds.length>=adminRoles.size()){//如果要删除的角色和该账户拥有的角色一样
+            throw new MyWebException("操作失败：账户至少要保留一个角色");
+        }
         Set<Integer> idsList=new HashSet<>();//要删除的id集合
         AdminRole adminRole=null;
         for(Integer roleId:roleIds){//遍历多个角色
@@ -207,6 +211,10 @@ public class AdminRoleServiceImpl implements AdminRoleService {
         if(adminId==null||roleId==null){
             throw new MyWebException("操作失败：非法参数");
         }
+        List<AdminRole> adminRoles=adminRoleDao.getAdminRoleByAdminId(adminId);//获取该角色的所有集合
+        if(adminRoles.size()==1){//如果要删除的角色和该账户拥有的角色一样
+            throw new MyWebException("操作失败：账户至少要保留一个角色");
+        }
         adminRoleDao.deleteByAdminIdAndRoleId(adminId,roleId);
     }
 
@@ -226,6 +234,11 @@ public class AdminRoleServiceImpl implements AdminRoleService {
         for(Integer adminId:adminIds){
             if(adminId==null){
                 throw new MyWebException("操作失败：非法参数（账户id为空）");
+            }
+            List<AdminRole> adminRoles=adminRoleDao.getAdminRoleByAdminId(adminId);//获取该角色的所有集合
+            if(roleIds.length>=adminRoles.size()){//如果要删除的角色和该账户拥有的角色一样
+                Admin admin=adminDao.get(adminId);
+                throw new MyWebException("操作失败：账户"+admin.getUsername()+"至少要保留一个角色");
             }
             for(Integer roleId:roleIds){//遍历多个角色
                 if(roleId==null){
@@ -259,6 +272,11 @@ public class AdminRoleServiceImpl implements AdminRoleService {
         }
         List<Integer> idsList=new ArrayList<Integer>();
         for(Integer adminId:adminIds){
+            List<AdminRole> adminRoles=adminRoleDao.getAdminRoleByAdminId(adminId);//获取该角色的所有集合
+            if(adminRoles.size()<=1){//如果账户只剩下不到一个角色
+                Admin admin=adminDao.get(adminId);
+                throw new MyWebException("操作失败：账户"+admin.getUsername()+"至少要保留一个角色");
+            }
             Object idObject=adminRoleDao.getId(adminId,roleId);
             if(idObject!=null){//如果能够找到相应记录
                 idsList.add((Integer)idObject);
